disable fallback font during artboard rendering
fixes #7479
The reason why the text looks different is that the editor uses a default font if it doesn't find a codepoint in the style defined font.
For example, in the font SF Pro, the symbol carriage return (\r) doesn't have a codepoint, but it does in the default font, which generates a new text run.
Share links, or our video recorder, don't have a default font so they skip that symbol.
This issue could happen with any other symbol that isn't found in the style font, so we decided to disable default fonts when rendering artboard text to guarantee they will look the same as at export time.

Diffs=
1adf3dbf4 disable fallback font during artboard rendering (#7480)