Fixing tool drawing
A lot of the tool draw methods aren't implemented with "render" for the Rive Renderer.

The Joystick, TextTool, and others weren't working. This PR cleans them up and removes a bunch of dependent methods which are no longer necessary. It also mostly gets the StateMachine stage to completely work with the Renderer abstraction (save for the icons which hack back to Canvas). When this part gets wrapped we can move to Rive Renderer for the state machine stage too.

~~This is in draft because I still have to get dashing working with RenderPath (which the state machine uses).~~

Diffs=
5c312867e Fixing tool drawing (#8028)