Fix layout size propagation
Adds a propagateSizeToChildren API to IntrinsicallySizeable that allows the IntrinsicallySizeable class itself to determine whether or not to pass its size down to children (for example, Shape needs to pass down to its ParametricPath but NslicedNode needs to early out).

NOTE: This is only currently required for the CPP runtime but added it to Dart as well in case we may use it in the future.

Diffs=
bb3b386c15 Fix layout size propagation (#8407)